Overhaul publications page
The same command is kept, but some creative liberties are taken to make
the output less verbose. In the context of publications, we only care
about venue and year. The title can be omitted, since this is more for
fun. We do keep the permissions to maintain the terminal immersion.